Create JSON: Update json data structure
Episodes are now called titles to be compatible with
other potential datasets, such as the MCU.
Include the series name in titles instead of as global
field, which allows us to differentiate between multiple
series within one json.
This commit is in preparation of combining all animated
series' into one json.