Create JSON: Get series from title id
Instead of passing the series via cli argument,
we now extract it from the episode's ID.
An episode's ID is unique and always starts
with a letter identical to the possible values
of the former argument.

This commit prepares to parse all csv files at
once to create a single json file with all
animated series episodes.